+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## 0.31.0
 
-* Dependency updates (Kafka 3.9.0, Vert.x 4.5.11, Netty 4.1.115.Final)
+* Dependency updates (Kafka 3.9.0, Vert.x 4.5.10)
+* Added support for creating a new topic via endpoint.
 
 ## 0.30.0